Added comment section for users in each task.
It's a lot, and I need to test it out w/ flask. I made an extra column for comments on each task, and used flask to limit the section to only show 50 chars. If you hit the view all button, it will pop up a modal that shows all of the comments. This is all using bootstrap formatting and elements. Same thing as before, used dummy flask variables for comments. We can remove or simplify if it doesn't work w/ flask. We can change the character limit if its too stretched out for what we want.