Added manager project view
Sorry for the inconsistencies in naming, I'll change it. Pretty similar to userprojectview.html. Just added the edit button for tasks, and the add button for adding tasks. Still need to do the comment section, I need to figure out if we want project-level comments, task-level comments or both. I'll atleast add the comment section so managers can view the user comments tomorrow.